Move rethinkdb-specific command tests to own file
